tweet-service/requirements.txt
flask 
python-dotenv 
psycopg2
datetime